+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.2 (2013-11-01)
+## 0.1.2 (2013-11-01)
 
 - http access (twisted instead of cherrypy)
 - javascript client added
@@ -6,12 +6,12 @@
 - logging service/http messages
 - interactive python client
 
-## 1.1 (2013-10-26)
+## 0.1.1 (2013-10-26)
 
 - travis integration (nosetests for python, phpunit for php)
 - new commands: clear, items, status
 
-## 1.0 (2013-10-21)
+## 0.1.0 (2013-10-21)
 
 - cache server
 - clients: python, php